\name{hotspotRM-package}
\alias{hotspotRM-package}
\alias{hotspotRM}
\docType{package}
\title{
\packageTitle{hotspotRM}
}
\description{
\packageDescription{hotspotRM}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{hotspotRM}

Maintainer: \packageMaintainer{hotspotRM}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = hotspotRM)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{hotspotRM}}
%\section{Documentation Index}{\packageIndices{hotspotRM}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
